Change 'Press SPACE' font size and color. Move setup code into setup.py.
